util/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
breadability==0.1.20
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
docopt==0.6.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl#sha256=c8ac64840c1eb3e3ca7bd38bd1e1c48fb0faeb2449d54d01d5ce629af4595775
Flask==2.3.2
fonttools==4.40.0
google==3.0.0
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.25.0
packaging==23.1
pandas==2.0.2
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.8
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pydantic==1.10.9
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sumy==0.11.0
thinc==8.1.10
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
Werkzeug==2.3.6
